libdtndht implemented
The pure C lib provides a simple usage of a bittorrent DHT to announce its own EID and Convergence Layer
Informations, needed to find a DTN Node. It also provides the search of DTN Nodes, announced on the
bittorrent DHT

The lib is in a very alpha status, a lot of features, as periodic announcing must be implemented

Also Group Support is not yet implemented.

All functions must be tested.